[refactor] 패키징
