+++ app/src/main/java/com/example/seoulgonggong/ui/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.example.seoulgonggong
+package com.example.seoulgonggong.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.seoulgonggong.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
