+++ app/src/main/res/layout/activity_public_sports_facility.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".facility.PublicSportsFacilityActivity">
+    tools:context=".ui.facility.PublicSportsFacilityActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/toolbar"
